just some verif changes

+++ push_swap/libft/ft_atoi.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nmouslim <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: nmouslim <nmouslim@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/03 16:15:11 by nmouslim          #+#    #+#             */
-/*   Updated: 2022/05/04 15:14:38 by nmouslim         ###   ########.fr       */
+/*   Updated: 2022/10/09 18:08:27 by nmouslim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_atoi(const char *nptr)
+long long int	ft_atoi(const char *nptr)
 {
-	long int	result;
-	int			minus;
-	int			i;
+	long long int	result;
+	int				minus;
+	int				i;
 
 	i = 0;
 	result = 0;
